feat: added proof-of-ledger zk example
