+++ examples/proof-of-ledger/.gitignore
@@ -1,0 +1,4 @@
+.DS_Store
+Cargo.lock
+methods/guest/Cargo.lock
+target/
