+++ examples/proof-of-ledger/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "host"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+methods = { path = "methods" }
+risc0-zkvm = { version = "1.2.3" }
+amaru-examples-shared = { path = "../shared" }
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+serde = "1.0"
