+++ examples/proof-of-ledger/methods/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "methods"
+version = "0.1.0"
+edition = "2021"
+
+[build-dependencies]
+risc0-build = { version = "1.2.3" }
+
+[package.metadata.risc0]
+methods = ["guest"]
