+++ examples/proof-of-ledger/methods/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    risc0_build::embed_methods();
+}
