+++ examples/proof-of-ledger/methods/guest/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "guest"
+version = "0.1.0"
+edition = "2021"
+
+[workspace]
+
+[dependencies]
+risc0-zkvm = { version = "1.2.3", default-features = false, features = ['std'] }
+amaru-examples-shared = { path = "../../../shared" }
