+++ examples/proof-of-ledger/methods/src/lib.rs
@@ -1,0 +1,1 @@
+include!(concat!(env!("OUT_DIR"), "/methods.rs"));
