+++ examples/proof-of-ledger/rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "1.83.0"
+components = ["rustfmt", "rust-src"]
+profile = "minimal"
