+++ examples/proof-of-ledger/src/main.rs
@@ -1,0 +1,54 @@
+// These constants represent the RISC-V ELF and the image ID generated by risc0-build.
+// The ELF is used for proving and the ID is used for verification.
+use methods::{
+    GUEST_ELF, GUEST_ID
+};
+use risc0_zkvm::{default_prover, ExecutorEnv};
+use amaru_examples_shared::RAW_BLOCK_CONWAY_1;
+
+fn main() {
+    // Initialize tracing. In order to view logs, run `RUST_LOG=info cargo run`
+    tracing_subscriber::fmt()
+        .with_env_filter(tracing_subscriber::filter::EnvFilter::from_default_env())
+        .init();
+
+    // An executor environment describes the configurations for the zkVM
+    // including program inputs.
+    // A default ExecutorEnv can be created like so:
+    // `let env = ExecutorEnv::builder().build().unwrap();`
+    // However, this `env` does not have any inputs.
+    //
+    // To add guest input to the executor environment, use
+    // ExecutorEnvBuilder::write().
+    // To access this method, you'll need to use ExecutorEnv::builder(), which
+    // creates an ExecutorEnvBuilder. When you're done adding input, call
+    // ExecutorEnvBuilder::build().
+
+    let input: String = RAW_BLOCK_CONWAY_1.to_string();
+    let env = ExecutorEnv::builder()
+        .write(&input)
+        .unwrap()
+        .build()
+        .unwrap();
+
+    // Obtain the default prover.
+    let prover = default_prover();
+
+    // Proof information by proving the specified ELF binary.
+    // This struct contains the receipt along with statistics about execution of the guest
+    let prove_info = prover
+        .prove(env, GUEST_ELF)
+        .unwrap();
+
+    // extract the receipt.
+    let receipt = prove_info.receipt;
+
+    // For example:
+    let _output: String = receipt.journal.decode().unwrap();
+
+    // The receipt was verified at the end of proving, but the below code is an
+    // example of how someone else could verify this receipt.
+    receipt
+        .verify(GUEST_ID)
+        .unwrap();
+}
