+++ examples/shared/assets/conway1.block
