+++ examples/shared/src/store.rs
